move report_urls generator as a utility
